fixed copy error
