+++ hoomd-interaction/src/cutoff_pair.rs
@@ -7,7 +7,7 @@
 use crate::{SitePairEnergy, TotalEnergy};
 use hoomd_microstate::{Microstate, property::Position};
 
-use hoomd_vector::Vector;
+use hoomd_vector::{Vector};
 
 /** Compute system properties given a [`SitePairEnergy`].
 
@@ -97,7 +97,7 @@ where
         pairwise::{Isotropic, LennardJones}};
     use hoomd_microstate::{Microstate, Body};
     use hoomd_microstate::property::{Point, Position};
-    use hoomd_vector::{Cartesian, Vector};
+    use hoomd_vector::{Cartesian, InnerProduct};
 
     # fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut microstate = Microstate::new();
