+++ hoomd-interaction/src/pairwise/isotropic.rs
@@ -7,7 +7,7 @@
 use super::IsotropicEnergy;
 use crate::SitePairEnergy;
 use hoomd_microstate::property::Position;
-use hoomd_vector::Vector;
+use hoomd_vector::{Vector};
 
 /** Compute isotropic properties from a pair of sites
 
@@ -49,7 +49,7 @@ where
 {
     #[inline]
     fn site_pair_energy(&self, a: &S, b: &S) -> f64 {
-        self.0.energy((*a.position() - *b.position()).norm())
+        self.0.energy((a.position()).distance(&(b.position())))
     }
 }
